+++ functions/pageSpeed.js
@@ -12,9 +12,13 @@ let bar;
  * @returns { Promise<void> } - return async mobile results
  */
 const mobile = async (url) => {
-  const res = await axios.get(
-    `https://www.googleapis.com/pagespeedonline/v5/runPagespeed?url=${url}&key=AIzaSyBEDaW4FxSZ2s1vz5CdD5Ai6PGZGdAzij0&strategy=mobile`
-  );
+  const res = await axios.get("https://www.googleapis.com/pagespeedonline/v5/runPagespeed", {
+    params: {
+      url,
+      key: "AIzaSyBEDaW4FxSZ2s1vz5CdD5Ai6PGZGdAzij0",
+      strategy: "mobile"
+    }
+  });
   
   try {
     const movil = res.data.lighthouseResult.categories.performance.score * 100;
@@ -73,9 +77,13 @@ const mobile = async (url) => {
  *
  */
 const desktop = async (url) => {
-  const res = await axios.get(
-    `https://www.googleapis.com/pagespeedonline/v5/runPagespeed?url=${url}&key=AIzaSyBEDaW4FxSZ2s1vz5CdD5Ai6PGZGdAzij0&strategy=desktop`
-  ); 
+  const res = await axios.get("https://www.googleapis.com/pagespeedonline/v5/runPagespeed", {
+    params: {
+      url,
+      key: "AIzaSyBEDaW4FxSZ2s1vz5CdD5Ai6PGZGdAzij0",
+      strategy: "desktop"
+    }
+  });
 
   try {
     const desktop = res.data.lighthouseResult.categories.performance.score * 100;
@@ -83,7 +91,7 @@ const desktop = async (url) => {
     switch (true) {
       case (desktop === 0 || desktop <=49):
         bar = new cliProgress.SingleBar({
-          format: `Mobile Result | ${colors.red("{bar}")} || {value}/{total} || bad`,
+          format: `Desktop Result | ${colors.red("{bar}")} || {value}/{total} || bad`,
           barCompleteChar: "\u2588",
           barIncompleteChar: "\u2591",
           hideCursor: true
@@ -91,7 +99,7 @@ const desktop = async (url) => {
         break;
       case (desktop === 50 || desktop <=89):
         bar = new cliProgress.SingleBar({
-          format: `Mobile Result | ${colors.yellow("{bar}")} || {value}/{total} decent`,
+          format: `Desktop Result | ${colors.yellow("{bar}")} || {value}/{total} decent`,
           barCompleteChar: "\u2588",
           barIncompleteChar: "\u2591",
           hideCursor: true
@@ -99,7 +107,7 @@ const desktop = async (url) => {
         break;
       case (desktop >=90 || desktop === 100):
         bar = new cliProgress.SingleBar({
-          format: `Mobile Result | ${colors.green("{bar}")} || {value}/{total} excelent`,
+          format: `Desktop Result | ${colors.green("{bar}")} || {value}/{total} excelent`,
           barCompleteChar: "\u2588",
           barIncompleteChar: "\u2591",
           hideCursor: true
@@ -107,7 +115,7 @@ const desktop = async (url) => {
         break;
       default:
         bar = new cliProgress.SingleBar({
-          format: `Mobile Result | ${colors.magenta("{bar}")} || {value}/{total} undifined`,
+          format: `Desktop Result | ${colors.magenta("{bar}")} || {value}/{total} undifined`,
           barCompleteChar: "\u2588",
           barIncomopleteChar: "\u2591",
           hideCursor: true