+++ test/index.test.js
@@ -1,7 +1,10 @@
+// @ts-nocheck
 // test modules
 const single = require("./single");
 const multiple = require("./multiple");
 const { desktop, mobile } = require("./pageSpeed");
+const githubInfo = require("./githubinfo");
+const animeSearch = require("./githubinfo");
 
 test('single stack', async () => await single("http://example.com"), 34000)
 
@@ -10,3 +13,7 @@ test('multiple stack', async () => await multiple(["http://example.com", "https:
 test('desktop speed', async () => await desktop("http://example.com"), 30000)
 
 test('mobile speed', async () => await mobile("http://example.com"), 30000)
+
+test('github info', async () => await githubInfo("faztweb"))
+
+test('anime search', async () => await animeSearch("naruto"))